requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.0
httpx==0.28.1
openai==1.58.1
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
python-telegram-bot==22.0
python-multipart==0.0.20
jinja2==3.1.5
pyyaml>=6.0.1
# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
# Redis
redis[hiredis]==5.2.1
# Task queue
arq==0.26.1
# Observability
sentry-sdk[fastapi]==2.19.2
prometheus-fastapi-instrumentator==7.0.2
structlog==24.4.0
# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
aiosqlite==0.20.0
